actually, don't couple the lastRunCommandNoFixit
